extract HasPhoneNumberConcern from User
